test(item): migrate counter
