+++ core/src/components/item/test/counter/item.e2e.ts-snapshots/item-counter-diff-ios-ltr-Mobile-Chrome-linux.png
