+++ core/src/components/item/test/counter/item.e2e.ts-snapshots/item-counter-diff-ios-rtl-Mobile-Safari-linux.png
